lean_requirements.txt
# Essential Dependencies for VeriPay Bot

# Telegram Bot Framework
aiogram==3.2.0

# Database
sqlalchemy==2.0.23

# Image Processing & OCR
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
pyzbar==0.1.9

# Configuration & Utilities
pyyaml==6.0.1

# Logging
loguru==0.7.2

# Basic utilities
requests==2.31.0
numpy==1.24.3 